fix project structure and delete unusable code
