+++ src/components/CompositeLayer/types.js
