--- src/components/Content/components/EditableLayerControl/index.css
@@ -1,31 +1,0 @@
-.editable-layer-control {
-  background-color: white;
-  border: 2px solid rgba(0, 0, 0, 0.2);
-  background-clip: padding-box;
-  box-sizing: content-box;
-  border-radius: 4px;
-}
-
-.editable-layer-control .content-row {
-  width: 100%;
-  height: 30px;
-  display: flex;
-  align-items: center;
-  padding-left: 8px;
-  padding-right: 8px;
-  box-shadow: 0 1px 0px 1px rgba(0.1, 0.1, 0.1, 0.2);
-}
-
-.editable-layer-control .content-row:last-child {
-  box-shadow: none;
-}
-
-.editable-layer-control .option:hover {
-  cursor: pointer;
-  background-color: rgba(0, 0, 0, 0.05);
-}
-
-.editable-layer-control input[type='radio'] {
-  margin-right: 3px;
-  cursor: pointer;
-}
