--- src/components/Content/index.css
@@ -1,4 +1,0 @@
-.content {
-  position: relative;
-  display: grid;
-}
