+++ src/components/EditableLayerService/helpers.js
