+++ src/components/MapBadge/index.jsx
