+++ src/components/StaticDataLoader/index.jsx
