+++ src/components/WithMapServices/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import StaticDataLoader from '../StaticDataLoader';
 
-import settings from '../../../../settings';
+import settings from '../../settings';
 
 const WithMapServices = Component => props => (
   <StaticDataLoader url={settings.urls.services}>
