+++ src/components/WithWebSocketConnection/index.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import io from 'socket.io-client';
 
-import WebSocketContext from '../../../../ws-context';
+import WebSocketContext from '../../ws-context';
 
 const WithWebSocketConnection = ({ services, children }) => {
   const [connections, setConnections] = useState(null);
